[lib] Pass threadType and parentThreadID to alwaysDisplayShortName ThreadEntity
Summary:
We need these params for `threadNoun` to be able to accurately pick a noun for thread.

We already pass `threadType` (when available) to `threadNoun`, so including that should "automatically" improve things.

`parentThreadID` is new, as is necessary to differentiate children of `GENESIS` from children of other communities. The following diff will update `threadNoun` to look at that.

Test Plan:
Tested in combination with the next two diffs:

1. Make sure that robotext for chat creation of GENESIS subchannels calls them "chats"
2. Make sure that robotext for chat creation of non-GENESIS subchannels calls them "channels"
3. Make sure that notifs for chat creation of GENESIS subchannels calls them "chats"
3. Make sure that notifs for chat creation of GENESIS subchannels doesn't mention GENESIS

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7078