[web] Add `CryptoStore` to `persistConfig.whitelist`
Summary:
We were previously excluding `CryptoStore` from `redux-persist` as we were working to avoid having to create a ton of migrations.

But now that we have primaryAccount/notificationAccount and primaryIdentityKeys/notificationIdentityKeys I think things are stable in the short term and we can start persisting.

---

Depends on D7081

Test Plan: Reload page and see that `cryptoStore` is included in `REHYDRATE` action:

Reviewers: ashoat, marcin, kamil

Reviewed By: ashoat

Subscribers: kamil, tomek

Differential Revision: https://phab.comm.dev/D7070