[native] add feature flag to avatars render work
Summary:
added `AVATARS_DISPLAY` feature flag to all avatars render work that is ready to land. For this diff created a new hook called `useShouldRenderAvatars` which determines if the current user has the
feature flag "on" and returns a boolean to either render or not render the avatar. The criteria for having the `AVATARS_DISPLAY` feature flag on is that they need to be a staff user. In subsequent diffs I
will add the feature flag when introducing things like the thread avatars and user avatars in chat screen.

Test Plan: Ginsu who is a staff user can see avatars, but Dan who is not a staff avatar can not see the rendered avatars on their screen

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7099