[native] Add a C++ MessageSpec for SidebarSource
Summary:
In D6869, I updated `SidebarSourceMessageSpec.generatesNotifs` so that it returned `pushTypes.NOTIF`. While `SidebarSourceMessageSpec.notificationTexts` should never be called, this still made it so `SidebarSource` messages could be included as the second item in notif payloads' `messageInfos`, since they get collapsed into the `CreateSidebar`.

As a result of being included in the notif payloads, `SidebarSource` is now getting processed by NSE, but since there is a missing C++ `MessageSpec` for it, we end up persisting an empty `content` to SQLite.

Test Plan: I was able to repro the "content must be defined" error before this diff by creating a sidebar that auto-added a test user. The test user would get the payload with `SidebarSource`, which would trigger the error. After this diff, the error no longer repros

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7085