[desktop] Stop checking for a new version after download
Summary:
This was the source of the issue [ENG-3311](https://linear.app/comm/issue/ENG-3311/), more information in the comments.
The short version is that if we checked for updates (and didn't find any) after downloading a new desktop version an error would appear when user tried to quit and update.

Test Plan:
Check if the error stops happening.

Reviewers: tomek, kamil

Subscribers: