[web/desktop] Update immediately on versions v1 and v2
Summary: Because of the issue explained in [ENG-3311](https://linear.app/comm/issue/ENG-3311/) and D7095, if user doesn't click the "Update" button on the modal right away, but instead waits ~10min or more they will get an error and won't be able to update. This diff makes it so if they are on one of the affected versions they will update immediately.

Test Plan:
- Check if drafts are persisted between updates
- Test if the forceful update goes through successfuly a few times, in different conditions like a laptop going to sleep

Reviewers: tomek, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7096