[lib][keyserver] Introduce new threadPermissions types for editing messages
Summary:
We sometimes would want to block editing messages, but still be able to write new messages to the thread.

This diff introduces the type, adds it to all roles except GENESIS members, and adds a new migration to populate it.

Test Plan: Checked if migration worked. The permissions check will be added with another diff.

Reviewers: ginsu, michal, inka

Subscribers: