[native] Introduce useNativeMessageListData and NativeChatMessageItem
Summary:
On `native`, we want to include information necessary for ENS resolution in this datatype, since it affects height calculations.

This diff splits the type out from the generic one we have in `lib`. For now the types are the same, but future diffs will change that.

Test Plan: I created [a patch](https://gist.github.com/Ashoat/02b230b25772395d1e0769ce073468fa) to add a field to `NativeChatMessageItem` and made sure it worked in Flow

Reviewers: ginsu, atul

Subscribers: