Build files set-up for Android 13
Summary:
This differential modifies Android build configuration files so that we can use Android 13 specific API on devices where it is
available. I was following instructions here: https://developer.android.com/develop/ui/views/notifications/notification-permission

Test Plan:
This differential should be tested on both Android 12 and Android 13 devices. For credible results it is recommended to used physical device rather than emulator.

Test on Android 13:
1. Fresh-install the app.
2. Make sure that no notifications permission prompt appears neither when logged in nor when logged out.
3. Make sure push notifs are disabled and not working unless it is directly changed in settings.
The reason for that is because since Android 13 notifications are not granted by default and unless app is configured to target Android 13 the system prompts for notifications permissions when it decides it is best to do so. If the app is configured to target Android 13, the system is not involved and it is the app responsibility to request those permissions. We have not implemented the code yet so permission are not granted until user directly grants them in settings app.

Test on Android 12:
We expected that older androids do not change their behaviour after applying this diffs as well as the entire stack. Push notifs are granted by default and should be working unless directly disabled in the settings app.

Both of those plans should be executed on dev and release builds.

Reviewers: tomek, bartek, jon, atul, ashoat

Reviewed By: tomek, bartek, atul, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7105