[Identity] Update X3DH protobuf to differiente key requests
Summary:
When establishing an X3DH communication channel, the
initiating device will need a onetime prekey, while the receiving
device will only need identity keys and prekeys. This change
is to allow for the prekey-only request to not also request
one-time keys.

https://linear.app/comm/issue/ENG-3345

Test Plan: N/A, this is mostly for design

Reviewers: varun, ashoat

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D7115