[keyserver] Display edited message in the sidebar if the source message is edited
Summary:
~~In case the user creates a sidebar from the message which has been edited, we need to add an edit message to a new thread to display updated content to the client.~~

In case the user creates a sidebar from the message which has been edited, we need to display updated content to the client. We are doing it by changing the content of `sourceMessage` of `SIDEBAR_SOURCE`.

Test Plan:
Create a new pending thread from the edited message. Send a message to it. A new thread is created, and the updated content of the first message is visible. The edited `source_message` is successfully sent to the client.
Create a new pending thread from an unedited message. Send a message and check if everything works as before.

Reviewers: michal, inka, ginsu, ashoat, rohan

Reviewed By: ashoat

Subscribers: rohan, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7141