[lib] Added edit message action reducer
Summary: After the edit action is completed, we receive from the keyserver new message infos. We can merge the new messages, to display edited messages locally.

Test Plan: Called the action by updating the message content. Checked if the messages are merged, and the edited messages updated with new content.

Reviewers: michal, inka, ginsu

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7144