[keyserver] Adding edit_message to the sidebar thread
Summary: This handles the situation, where the user edits a message which has a sidebar. It adds edit_message to the sidebar thread.

Test Plan: Called editing message endpoint for a message which has a sidebar. The edit message has been added to both the main thread and the sidebar thread. Reducer on client side properly
merges two new messages instead of one.

Reviewers: michal, inka, ginsu

Subscribers: