[keyserver/lib] fetch avatar column from users table in fetchCurrentUserInfo
Summary:
The current user will need the avatar info from the db. Updated the query to include `avatar` in the returned object whenever `avatar` is not null. A subsequent diff will address the other queries.

**This diff won't be landed until there is a feature flag for the avatars work**

---

Linear Task: https://linear.app/comm/issue/ENG-3209/update-user-fetchers-to-return-avatar-info-from-users-table

Test Plan:
I logged out `currentUserInfo` in `profile-screen.react`

When `avatar` is null:
{F427091}

When `avatar` is populated (hardcoded an avatar variable on the keyserver just for the logs)
{F427092}

Reviewers: ashoat, atul

Reviewed By: ashoat, atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7054