[web] Add grpc-web@1.4.2
Summary: We need this to talk to Comm services.

Test Plan:
I haven't really tested using this library. I did make sure `yarn cleaninstall` worked.

I worked on this with @jon in a 1:1, and I think the plan is for him to introduce usage once we have the `opaque-ke` WASM bindings working.

Reviewers: jon, varun, atul

Reviewed By: varun

Subscribers: tomek, jon

Differential Revision: https://phab.comm.dev/D7149