[server] Update assetlinks file
Summary:
We have to update the file so that we can deeplink to our app on Android. Figuring out what to put there is a lot easier than for iOS as the whole configuration is prepared by Google Play Console and can be copied from there.

Test Plan:
After the file is deployed to the server, I'll check if we're able to open the app through the link. I've checked using json validator if a file is properly formatted. Also checked if the file can be accessed at `/.well-known/assetlinks.json`.

Reviewers:
bartek, kamil