[keyserver/lib] Introduce a new message type and spec for toggling pins
Summary: Part of the way we want to alert clients that a message has been pinned / unpinned is by introducing a new message type and message spec to represent toggling a pin. I've decided to consolidate the
two actions into one spec and handle the robotext / notif texts that are sent within the functions to avoid lots of duplicate code (much like `update-relationship-message-spec.js`).

Depends on D7112

Linear: https://linear.app/comm/issue/ENG-3190/introduce-a-new-message-type-and-spec-for-toggling-pins

Test Plan:
1. Confirmed that `hasMinCodeVersion` successfully works for both cases where the client should be able to see the new message type and vice versa.
2. Ran several pin / unpin actions on a variety of messages (text, images, videos) to confirm that the robotext is as expected

Reviewers: atul, ginsu, tomek

Subscribers: