[keyserver] `JSON.parse()` stringified `avatar` in `fetchLoggedInUserInfo`
Summary:
We want to `JSON.parse()` the stringified `avatar` from the database before including in `LoggedInUserInfo` and returning.

Noticed that on user login `avatar` was a string instead of an object of type `ClientAvatar`.

The `mysql2` node package conveniently "auto parsed" stringified JSON when providing the result set with `MySQL`, but unfortunately doesn't for MariaDB (https://github.com/sidorares/node-mysql2/issues/1287)... so we need to `JSON.parse(...)` ourself.

Test Plan:
Logged in and `avatar` was object of type `ClientAvatar` as expected.

Reviewers:
ashoat, tomek