[lib][keyserver] Attach MIME and encryption key to uploads
Summary:
Part of [[ https://linear.app/comm/issue/ENG-3393/keyserver-support-for-encrypted-messages | ENG-3393 ]]

In order to upload encrypted media, we need to store encryption key on the keyserver.
Also, MIME type can no longer be detected on keyserver side, so we need to send it along with the media

This diff modifies the `upload_multimedia` endpoint to support providing these values. They're validated, but not used anywhere yet.

MIME type is already available when uploading, it only needed to be exposed.

Test Plan: Modified native input-state-container to provide example encryption key. Added console.log on keyserver side to see if both values are passed correctly.

Reviewers: atul, marcin, ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7166