[keyserver] Process and store encrypted media
Summary:
Part of [[ https://linear.app/comm/issue/ENG-3393/keyserver-support-for-encrypted-messages | ENG-3393 ]]

Encrypted media can no longer be processed on keyserver. Modified `validateAndConvert` - added a simple validation based on provided MIME type - it must be provided in order to detect if we have image or video. Then the content is passed through.

Then stored the encryption key in `extras` column of the `uploads` table.

Depends on D7166

Test Plan: Modified the `uploadMultimedia` call in input-state-container to pass some encryptionKey. Uploaded some media. Ensured the `encryptionKey` is added to the `uploads` table

Reviewers: atul, marcin, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7167