Allow while(true) in the JS codebase
Summary: I find that it's sometimes better for readability for the condition of a loop to be determined by a `break` call inside the body of the loop. More context [here](https://phab.comm.dev/D7117?id=23920#inline-46985).

Test Plan: I edited a file and added a `while (true) { break; }` line. Before this diff it causes an ESLint error, but afterwards it didn't

Reviewers: inka, atul, tomek, bartek

Reviewed By: atul

Differential Revision: https://phab.comm.dev/D7173