[web-db] add table and operations for redux-persist engine
Summary:
Supporting operations for custom [storage-engine](https://github.com/rt2zz/redux-persist#storage-engines)

NOTE: There is a code repetition between this and metadata because both are basically operations on a key-value table. Creating a [task](https://linear.app/comm/issue/ENG-3406/unify-queries-for-key-value-tables) to unify this as a follow-up.

NOTE: D7121 is not landed yet, and even so, my local environment is the only place where this code was executed so adding this without migration.

Depends on D7182

Test Plan: Run tests D7184

Reviewers: michal, tomek, marcin, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7183