[keyserver] Don't gate `avatar` inclusion in `*LoggedInUserInfo` based on `codeVersion`
Summary:
After some discussion we clarified that we do want to include the `avatar` field for old clients IFF it's set (ie not `null`).

I mistakenly suggested to @ginsu that we include this check when reviewing D7054 which was my bad.

Test Plan: Avatar should be included in `*UserInfo` even for older clients.

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: tomek, ginsu

Differential Revision: https://phab.comm.dev/D7176