[keyserver] Convert user ID to VARCHAR in MariaDB
Summary:
Identity service is going to own user IDs going forward, and they are no longer going to be exclusively numeric.

To pave the way for this, we're converting all user ID columns to `VARCHAR(255)` in the `latin1` charset.

Linear task: https://linear.app/comm/issue/ENG-3279/migrate-mariadb-user-ids-to-varchar-type

Test Plan:
1. I tested by migrating an existing database and creating a new one
2. I tested creating a new user
3. I tested logging in an existing user
4. I tested sending messages
5. I tested sending / accepting a friend request
6. I tested unfriending
7. I tested blocking / unblocking

Reviewers: atul, jon, tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D7179