[web] render user avatars in chat screen
Summary:
Rendered the user avatars in the chat screen. An important thing to note about this diff vs D7067 is that since web has a lot more screen real estate than native, I DID NOT reduce the width of chat bubbles. I thought about it at first; however, after seeing how the chat screen looked visually with the avatar, I felt that it was unnecessary to reduce the width of the chat bubbles as we did in native

Depends on D7194

Test Plan:
Please look at the screenshots below to see the changes I made (only handles the user avatars in the chat screen and none of the thread avatars shown in the UI--- those will come later in the stack)

Before:
{F448367}

After:
{F448368}

Under the hood:

Before:
{F448291}

After:
{F448292}

Responsiveness:

Before:
{F448293}

After:
{F448300}

Responsiveness with super thin photos:

Before:
{F448314}

After:
{F448309}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7195