[keyserver] Introduce endpoint for deep linking
Summary: This endpoint won't be accessed when the app is installed - opening it on a mobile device should deep link to the app. If the app isn't installed, this endpoint will be called on the server side. Then we can redirect the user to the store. At this point only the App Store is supported but later in the stack Google Play will be also handled - based on the user agent. This is also the place where deferred part of linking will happen - setting referrer or clipboard item.

Test Plan: Create a test page containing a link to `/invite/:secret`. Clicking this link on a device with installed app results in opening the app with appropriate `initialURL` set. Clicking it on other devices opens the App Store.

Reviewers: bartek, kamil

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7217