[web-db] support SQLite for staff and non-mobile browsers
Summary:
Adding flag in website responder to get this information at the very begging of web app lifecycle.

I did some research and there is no one good method for detecting it is mobile browser. I choose solution which I think is most likely to work, but it'll be good to revisit this in the future [ENG-3407](https://linear.app/comm/issue/ENG-3407/revisit-method-for-detecting-mobile-browser).

Test Plan:
1. Test for web app and electron
2. Test in browser on mobile

Reviewers: michal, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7188