[native] render thread avatars in community drawer
Summary:
rendered thread avatar in community drawer. One thing to note is that I did notice that the font size for the thread name was getting decreased the deeper it goes into the community tree. I asked Ted last week if we should do the same with Avatars and he said no. He also said that it would be nice to eventually have the font size for the thread name be the same regardless of level in the community tree. I figured that making the font size all the same might require more discussion/is outside the scope of this diff

---

Depends on D7135

Test Plan:
Please check out these screenshots to see the changes I made:

Staff user view
{F445626}

non staff user view:
{F445627}

Reviewers: ashoat, atul, inka

Reviewed By: ashoat

Subscribers: inka, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7136