[keyserver] Include debug fingerprint
Summary: Initially this file included only a fingerprint downloaded from Google Play Console - which is necessary for the released app, but doesn't work for local development. This new fingerprint was generated using Android Studio and is dedicated for debug configuration. It is unlikely that this fingerprint will work for other devs - the right solution is to share debug keystore and use it for the assets file. I'm going to create a task for the right solution, but this workaround should help me while working on the deep links.

Test Plan: Checked if this file is a correct JSON. Further testing can be performed after it is deployed - it should allow deeplinking to debug app.

Reviewers: bartek, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7219