[native] Enable App Link
Summary: Open the app when a link is opened.

Test Plan: Tested it by creating a page with a link to `https://web.comm.app/intive/123` - clicking it opens the app when it is closed and when it is in the background. Also checked opening it using adb activity manager.

Reviewers: bartek, marcin, kamil

Reviewed By: bartek

Subscribers: marcin, ashoat, atul

Differential Revision: https://phab.comm.dev/D7255