[lib] Add option to force-create multimedia-type message
Summary:
Currently, when creating a new multimedia message, it's type is based on content. If there are only photos, then it is set to `PHOTOS` (14). If it contains one or more videos, then it is set to `MULTIMEDIA` (15).
For encrypted photos, the type should always be `MULTIMEDIA`. This diff adds an option to force this setting.

Test Plan: Set this option to `true` in the callsite inside input-state-container. Then sent a message. Ensured that this message has type `15` in both Redux and keyserver database.

Reviewers: ashoat, atul, marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7230